Split test_binaries
