+++ docs/source/tests/python/compilation.rst
@@ -2,3 +2,5 @@ Compilation
 ===========
 
 .. automodule:: tests.python.compilation
+
+.. automodule:: tests.python.test_compilation
