+++ docs/source/tests/python/tools/binaries/cuobjdump.rst
@@ -1,0 +1,4 @@
+cuobjdump
+=========
+
+.. automodule:: tests.python.tools.binaries.test_cuobjdump
