+++ docs/source/tests/python/tools/binaries/demangle.rst
@@ -1,0 +1,4 @@
+demangle
+========
+
+.. automodule:: tests.python.tools.binaries.test_demangle
