+++ tests/python/tools/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_pytest(architecture)
-add_pytest(binaries)
 add_pytest(cacher)
 add_pytest(device_properties)
 add_pytest(ncu)
