+++ tests/python/tools/binaries/CMakeLists.txt
@@ -1,1 +1,3 @@
+add_pytest(cuobjdump)
+add_pytest(demangle)
 add_pytest(elf)
